# Performance improvements using a cursor instead of a shift