Pull in performance improvement for large row counts.

Closes: #56 

I tested this first with 100x run of with fruits.parquet (40,000 rows) and at least with Node 14+ and Node 16+ it is faster with the `shift` version by ~3% (old: ~15s, new ~15.5s).

Then I tried again (1x run) with a file that was 7,667,792 rows (https://ursa-labs-taxi-data.s3.us-east-2.amazonaws.com/2019/01/data.parquet):
shift: ~190s
cursor: ~34s

Major difference.